finalize auth service

+++ src/components/moviesTable.jsx
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import auth from "../services/authService";
 import { Link } from "react-router-dom";
 import Table from "./common/table";
 import Like from "./common/like";
@@ -23,16 +24,22 @@ class MoviesTable extends Component {
     },
     {
       key: "delete",
-      content: (movie) => (
+      content: (movie) => this.renderDeleteContent(movie),
+    },
+  ];
+
+  renderDeleteContent(movie) {
+    const user = auth.getCurrentUser();
+    if (user && user.isAdmin)
+      return (
         <button
           onClick={() => this.props.onDelete(movie)}
           className="btn btn-danger btn-sm"
         >
           Delete
         </button>
-      ),
-    },
-  ];
+      );
+  }
 
   render() {
     const { movies, onSort, sortColumn } = this.props;
